Again, renaming router due to http://stackoverflow.com/questions/19341975/heroku-node-cannot-find-module-error
